fix: responsivity

+++ frontend/src/app/components/students/students.component.ts
@@ -4,17 +4,20 @@ import { Student } from '../../model/Student';
 import { CommonModule } from '@angular/common';
 import { MatIconModule } from '@angular/material/icon';
 import { differenceInYears, parseISO } from 'date-fns';
+import { FormsModule } from '@angular/forms';
 
 @Component({
   selector: 'app-students',
   standalone: true,
-  imports: [CommonModule, MatIconModule],
+  imports: [CommonModule, MatIconModule, FormsModule],
   templateUrl: './students.component.html',
   styleUrl: './students.component.scss'
 })
 export class StudentsComponent implements OnInit {
 
   students: Student[] = [];
+
+  searchByName: string = '';
 
   constructor(private studentsService: StudentsService) { }
 
@@ -42,13 +45,21 @@ export class StudentsComponent implements OnInit {
   }
 
   deleteStudent(studentId: number ): void {
-    this.studentsService.deleteStudent(studentId).subscribe(() => {
-      this.students = this.students.filter(student => student.studentId !== studentId)
-    });
+    if (confirm('Tem certeza de que deseja deletar este aluno?')) {
+      this.studentsService.deleteStudent(studentId).subscribe(() => {
+        this.students = this.students.filter(student => student.studentId !== studentId);
+      });
+    }  
   }
 
   handleShowGrade(student: Student): void {
     student.showGrade = !student.showGrade;
+  }
+
+  getStudentsByName() {
+    return this.students.filter(student => 
+      student.name?.toLowerCase().includes(this.searchByName.toLowerCase())
+    );
   }
 
 }
